backend/requirement.txt
bcrypt
beautifulsoup4
deep-translator
faster-whisper
Flask
Flask-Cors
Flask-Mail
Flask-Session
google-auth
google-auth-oauthlib
google-generativeai
mysql-connector-python
pdfplumber
Pillow
pydub
PyJWT
PyPDF2
python-dotenv
pytesseract
pytube
requests
SpeechRecognition
torch
transformers
youtube-transcript-api
